+++ packages/@expo/fingerprint/src/sourcer/__tests__/Expo-test.ts
@@ -538,8 +538,8 @@ describe(`getExpoConfigSourcesAsync - sourceSkips`, () => {
       vol.fromJSON(require('./fixtures/ExpoManaged47Project.json'));
       mockConfigFile('/app/app.config.js', () => ({
         default: ({ config }: any) => {
-          config.android = { versionCode: 1 };
-          config.ios = { buildNumber: '1' };
+          config.android = { versionCode: 1, version: '1.2.3' };
+          config.ios = { buildNumber: '1', version: '4.5.6' };
           return config;
         },
       }));
@@ -557,34 +557,9 @@ describe(`getExpoConfigSourcesAsync - sourceSkips`, () => {
       expect(expoConfig).not.toBeNull();
       expect(expoConfig.version).toBeUndefined();
       expect(expoConfig.android.versionCode).toBeUndefined();
-      expect(expoConfig.ios.buildNumber).toBeUndefined();
-    });
-  });
-
-  it('should remove ios.version and android.version when SourceSkips.ExpoConfigPlatformVersions', async () => {
-    await jest.isolateModulesAsync(async () => {
-      vol.fromJSON(require('./fixtures/ExpoManaged47Project.json'));
-      mockConfigFile('/app/app.config.js', () => ({
-        default: ({ config }: any) => {
-          config.android = { version: '1.2.3' };
-          config.ios = { version: '4.5.6' };
-          return config;
-        },
-      }));
-
-      const options = await normalizeOptionsAsync('/app', {
-        sourceSkips: SourceSkips.ExpoConfigPlatformVersions,
-      });
-      const { config, loadedModules } = await getExpoConfigAsync('/app', options);
-      const sources = await getExpoConfigSourcesAsync('/app', config, loadedModules, options);
-      const expoConfigSource = sources.find<HashSourceContents>(
-        (source): source is HashSourceContents =>
-          source.type === 'contents' && source.id === 'expoConfig'
-      );
-      const expoConfig = JSON.parse(expoConfigSource?.contents?.toString() ?? 'null');
-      expect(expoConfig).not.toBeNull();
-      expect(expoConfig.ios.version).toBeUndefined();
       expect(expoConfig.android.version).toBeUndefined();
+      expect(expoConfig.ios.buildNumber).toBeUndefined();
+      expect(expoConfig.ios.version).toBeUndefined();
     });
   });